+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/composable/actions/ComposableActionFlipper.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.keyscreen.impl.composable.actions
 
-import android.annotation.SuppressLint
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
@@ -10,6 +9,7 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -31,12 +31,13 @@ import androidx.compose.ui.unit.sp
 import com.flipperdevices.core.ui.R as DesignSystem
 
 @Composable
+@Suppress("LongParameterList")
 fun ComposableActionFlipper(
+    modifier: Modifier = Modifier,
     @DrawableRes iconId: Int,
     @StringRes descriptionId: Int,
-    @ColorRes descriptionColorId: Int = com.flipperdevices.core.ui.R.color.white_100,
-    @ColorRes tintId: Int = com.flipperdevices.core.ui.R.color.white_100,
-    @SuppressLint("ModifierParameter") modifier: Modifier = Modifier,
+    @ColorRes descriptionColorId: Int = DesignSystem.color.white_100,
+    @ColorRes tintId: Int = DesignSystem.color.white_100,
     onClick: () -> Unit
 ) {
     val descriptionText = stringResource(descriptionId)
@@ -49,7 +50,8 @@ fun ComposableActionFlipper(
                 interactionSource = remember { MutableInteractionSource() },
                 indication = rememberRipple(),
                 onClick = onClick
-            ).then(modifier)
+            )
+            .then(modifier)
     ) {
         Column(
             modifier = Modifier
@@ -58,18 +60,72 @@ fun ComposableActionFlipper(
             verticalArrangement = Arrangement.Center,
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Icon(
-                modifier = Modifier.size(size = 36.dp),
-                painter = painterResource(iconId),
-                contentDescription = descriptionText,
-                tint = colorResource(tintId)
-            )
-            Text(
-                text = descriptionText,
-                fontWeight = FontWeight.W500,
-                fontSize = 16.sp,
-                color = colorResource(descriptionColorId)
-            )
+            ComposableActionFlipperContent(iconId, descriptionText, descriptionColorId, tintId)
         }
     }
+}
+
+@Composable
+@Suppress("LongParameterList")
+fun ComposableActionFlipperHorizontal(
+    modifier: Modifier = Modifier,
+    @DrawableRes iconId: Int,
+    @StringRes descriptionId: Int,
+    @ColorRes descriptionColorId: Int = DesignSystem.color.white_100,
+    @ColorRes tintId: Int = DesignSystem.color.white_100,
+    onClick: (() -> Unit)? = null
+) {
+    val descriptionText = stringResource(descriptionId)
+
+    var boxModifier = modifier
+        .clip(shape = RoundedCornerShape(30.dp))
+
+    if (onClick != null) {
+        boxModifier = boxModifier
+            .background(colorResource(id = DesignSystem.color.accent_secondary))
+            .clickable(
+                interactionSource = remember { MutableInteractionSource() },
+                indication = rememberRipple(),
+                onClick = onClick
+            )
+    } else {
+        boxModifier = boxModifier
+            .background(colorResource(id = DesignSystem.color.black_40))
+    }
+
+    Box(
+        modifier = boxModifier,
+        contentAlignment = Alignment.Center
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(vertical = 12.dp),
+            horizontalArrangement = Arrangement.Center,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            ComposableActionFlipperContent(iconId, descriptionText, descriptionColorId, tintId)
+        }
+    }
+}
+
+@Composable
+private fun ComposableActionFlipperContent(
+    @DrawableRes iconId: Int,
+    descriptionText: String,
+    @ColorRes descriptionColorId: Int,
+    @ColorRes tintId: Int
+) {
+    Icon(
+        modifier = Modifier.size(size = 36.dp),
+        painter = painterResource(iconId),
+        contentDescription = descriptionText,
+        tint = colorResource(tintId)
+    )
+    Text(
+        text = descriptionText,
+        fontWeight = FontWeight.W500,
+        fontSize = 16.sp,
+        color = colorResource(descriptionColorId)
+    )
 }