+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - [BUGFIX] Fix reconnect issues
 - [BUGFIX] Fix synchronization start after first login
 - [Feature] Add report bug button
-- [Feature] Write and emulate buttons view
+- [Feature] Emulate button on key screen
 - [Feature] Add update Flipper via OTA on separate screen
 - [Feature] Added an option to cancel existing Flipper RPC requests
 - [Feature] Use new design for unsupported card and for unsupported dialog